+++ TD2/src/tri.hpp
@@ -13,7 +13,7 @@ void merge_sort(std::vector<float> & vec, size_t const left, size_t const right)
 void merge_sort(std::vector<float> & vec);
 
 //tableau aléatoire
-std::vector<int> generate_random_vector(size_t const size, int const max = 100);
+std::vector<float> generate_random_vector(size_t const size, int const max = 100);
 
 //dichotomie
 int search(std::vector<int> vec, int entier);
